+++ CHANGELOG.md
@@ -7,9 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `scalr_workspace`: added new attribute `execution-mode` ([#158](https://github.com/Scalr/terraform-provider-scalr/pull/158))
+- `data.scalr_workspace`: added new attribute `execution-mode` ([#158](https://github.com/Scalr/terraform-provider-scalr/pull/158))
+
+### Deprecated
+
+- `scalr_workspace`: attribute `operations` has been deprecated ([#158](https://github.com/Scalr/terraform-provider-scalr/pull/158))
+
+### Required
+
+- scalr-server >= `8.29.0`
+
 ## [1.0.0-rc33] - 2022-07-22
 
 ### Added
+
 - `scalr_workspace`: added new attribute `vcs_repo.ingress_submodules` ([#146](https://github.com/Scalr/terraform-provider-scalr/pull/146))
 - `data.scalr_workspace`: added new attribute `vcs_repo.ingress_submodules` ([#146](https://github.com/Scalr/terraform-provider-scalr/pull/146))